Added warning for old style cast

+++ test/data_structure/fenwick_tree_test.cpp
@@ -18,9 +18,10 @@ TEST(fenwick_tree, IsZeroConstructed) {
 
 TEST(fenwick_tree, SumPrefixAndIsMutable) {
   djp::fenwick_tree<long> ftree(10);
-  for (size_t i = 0; i < ftree.size(); ++i) ftree.increase(i, 10 * i);
+  for (size_t i = 0; i < ftree.size(); ++i)
+    ftree.increase(i, 10 * i);
   for (size_t i = 0, acc = 0; i < ftree.size(); ++i, acc += i * 10)
-    EXPECT_EQ((long)acc, ftree.sum(i));
+    EXPECT_EQ(acc, ftree.sum(i));
 
   EXPECT_EQ(80 + 90, ftree.sum(8, 9));
   EXPECT_EQ(30 + 40 + 50, ftree.sum(3, 5));
